requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter

dvc
mlflow

fastapi
uvicorn

streamlit

python-dotenv
pyyaml